requirements.txt
hazelcast-python-client==5.3.0
loguru==0.7.2
PyYAML==6.0.1
